+++ core/Exception/Exception.php
@@ -6,25 +6,25 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  *
  */
-namespace Piwik\Exceptions;
+namespace Piwik\Exception;
 
-use Exception;
+use Exception as PhpException;
 
 /**
  * An exception whose message has HTML content. When these exceptions are caught
  * the message will not be sanitized before being displayed to the user.
- *
- * @api
  */
-class HtmlMessageException extends Exception
+abstract class Exception extends PhpException
 {
-    /**
-     * Returns the exception message.
-     *
-     * @return string
-     */
-    public function getHtmlMessage()
+    private $isHtmlMessage = false;
+
+    public function setIsHtmlMessage()
     {
-        return $this->getMessage();
+        $this->isHtmlMessage = true;
+    }
+
+    public function isHtmlMessage()
+    {
+        return $this->isHtmlMessage;
     }
 }