+++ plugins/ScheduledReports/tests/Integration/ApiTest.php
@@ -19,8 +19,8 @@ use Piwik\ScheduledTask;
 use Piwik\ScheduledTime\Monthly;
 use Piwik\ScheduledTime;
 use Piwik\Site;
+use Piwik\Tests\Framework\Mock\FakeAccess;
 use Piwik\Tests\Framework\TestCase\IntegrationTestCase;
-use FakeAccess;
 use Exception;
 use ReflectionMethod;