+++ plugins/UsersManager/tests/Integration/APITest.php
@@ -8,10 +8,10 @@
 
 namespace Piwik\Plugins\UsersManager\tests;
 use Piwik\Access;
-use FakeAccess;
 use Piwik\Piwik;
 use Piwik\Plugins\UsersManager\API;
 use Piwik\Tests\Framework\Fixture;
+use Piwik\Tests\Framework\Mock\FakeAccess;
 use Piwik\Tests\Framework\TestCase\IntegrationTestCase;
 
 /**