+++ tests/resources/TestPluginLogClass.php
@@ -5,6 +5,7 @@
  * @link http://piwik.org
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
+
 namespace Piwik\Plugins\TestPlugin;
 
 use Piwik\Log;